YT-17615: Support core/net & core/http for windows Switch allocator in logging tests Fix crash on logging with null default log manager Use send and recv on windows Fix of propagate storage callbacks and disable logging test on windows Fix oom tests Use read/write instead of recv/send Net & http fixes for windows

+++ yt/yt/core/rpc/grpc/server.cpp
@@ -714,7 +714,7 @@ private:
             }
 
             auto micros = gpr_timespec_to_micros(gpr_time_sub(deadline, now));
-            if (micros > std::numeric_limits<ui64>::max() / 2) {
+            if (micros > static_cast<double>(std::numeric_limits<ui64>::max() / 2)) {
                 return;
             }
